Extract tests for isInt8Array to own file
